+++ yazi-fm/src/app/commands/bootstrap.rs
@@ -3,7 +3,7 @@ use yazi_actor::Ctx;
 use yazi_boot::BOOT;
 use yazi_macro::act;
 use yazi_parser::{VoidOpt, mgr::CdSource};
-use yazi_shared::{data::Data, url::UrlLike};
+use yazi_shared::{data::Data, path::PathBufLike, url::UrlLike};
 
 use crate::app::App;
 
@@ -18,7 +18,7 @@ impl App {
 			let cx = &mut Ctx::active(&mut self.core);
 			cx.tab = i;
 
-			if file.as_os_str().is_empty() {
+			if file.is_empty() {
 				act!(mgr:cd, cx, (BOOT.cwds[i].clone(), CdSource::Tab))?;
 			} else {
 				act!(mgr:reveal, cx, (BOOT.cwds[i].join(file), CdSource::Tab))?;