+++ yazi-parser/src/mgr/hover.rs
@@ -1,14 +1,13 @@
-use std::path::PathBuf;
-
 use mlua::{ExternalError, FromLua, IntoLua, Lua, Value};
+use yazi_shared::path::PathBufDyn;
 
 #[derive(Debug, Default)]
 pub struct HoverOpt {
-	pub urn: Option<PathBuf>,
+	pub urn: Option<PathBufDyn>,
 }
 
-impl From<Option<PathBuf>> for HoverOpt {
-	fn from(urn: Option<PathBuf>) -> Self { Self { urn } }
+impl From<Option<PathBufDyn>> for HoverOpt {
+	fn from(urn: Option<PathBufDyn>) -> Self { Self { urn } }
 }
 
 impl FromLua for HoverOpt {