+++ yazi-shared/src/path/buf.rs
@@ -4,11 +4,13 @@ use crate::path::{AsInnerView, AsPathView, PathInner, PathLike};
 
 pub trait PathBufLike
 where
-	Self: Default + 'static,
+	Self: 'static,
 {
 	type Inner: for<'a> AsInnerView<'a, Self::InnerRef<'a>>;
 	type InnerRef<'a>: PathInner<'a>;
 	type Borrowed<'a>: PathLike<'a> + AsPathView<'a, Self::Borrowed<'a>> + Debug + Hash;
+
+	fn borrow(&self) -> Self::Borrowed<'_>;
 
 	fn encoded_bytes(&self) -> &[u8];
 
@@ -16,17 +18,23 @@ where
 
 	fn into_encoded_bytes(self) -> Vec<u8>;
 
+	fn is_empty(&self) -> bool { self.encoded_bytes().is_empty() }
+
 	fn len(&self) -> usize { self.encoded_bytes().len() }
 
 	fn set_file_name<T>(&mut self, name: T)
 	where
 		T: for<'a> AsInnerView<'a, Self::InnerRef<'a>>;
+
+	fn take(&mut self) -> Self;
 }
 
 impl PathBufLike for std::path::PathBuf {
 	type Borrowed<'a> = &'a std::path::Path;
 	type Inner = std::ffi::OsString;
 	type InnerRef<'a> = &'a std::ffi::OsStr;
+
+	fn borrow(&self) -> Self::Borrowed<'_> { self.as_path() }
 
 	fn encoded_bytes(&self) -> &[u8] { self.as_os_str().as_encoded_bytes() }
 
@@ -42,4 +50,6 @@ impl PathBufLike for std::path::PathBuf {
 	{
 		self.set_file_name(name.as_inner_view());
 	}
+
+	fn take(&mut self) -> Self { std::mem::take(self) }
 }